test: switch to ava
