+++ .circleci/config.yml
@@ -1,7 +1,7 @@
 version: 2.1
 orbs:
   node: circleci/node@4
-  codecov: codecov/codecov@1
+  codecov: codecov/codecov@3
 jobs:
   test:
     executor:
@@ -29,7 +29,6 @@ jobs:
 workflows:
   version: 2
   test_and_release:
-    # Run the test jobs first, then the release only when all the test jobs are successful
     jobs:
       - test
       - release:
