+++ demo/package.json
@@ -8,10 +8,10 @@
   },
   "devDependencies": {
     "@tailwindcss/forms": "0.3.3",
-    "autoprefixer": "10.3.2",
-    "postcss": "8.3.6",
-    "tailwindcss": "2.2.7",
-    "typescript": "4.3.5",
-    "vite": "2.5.0"
+    "autoprefixer": "10.3.6",
+    "postcss": "8.3.8",
+    "tailwindcss": "2.2.16",
+    "typescript": "4.4.3",
+    "vite": "2.5.10"
   }
 }
