+++ package.json
@@ -26,24 +26,33 @@
     "lint:fix": "eslint --fix --ext .ts .",
     "prepare": "npm run build",
     "build": "tsc",
-    "test": "NODE_OPTIONS=\"--loader ts-node/esm\" uvu test",
-    "test:watch": "NODE_OPTIONS=\"--loader ts-node/esm\" watchlist src test -- uvu test",
-    "test:ci": "NODE_OPTIONS=\"--loader ts-node/esm\" c8 --reporter=text --reporter=lcov uvu test"
+    "test": "ava",
+    "test:watch": "ava --watch",
+    "test:ci": "c8 --reporter=text --reporter=lcov ava"
   },
   "dependencies": {
     "@ctrl/ts-base32": "^2.1.1"
   },
   "devDependencies": {
-    "@ctrl/eslint-config": "2.0.13",
-    "@sindresorhus/tsconfig": "1.0.2",
-    "@types/node": "16.7.1",
+    "@ctrl/eslint-config": "2.1.0",
+    "@sindresorhus/tsconfig": "2.0.0",
+    "@types/node": "16.10.1",
+    "ava": "4.0.0-alpha.2",
     "buffer": "6.0.3",
-    "c8": "7.8.0",
-    "expect": "27.0.6",
+    "c8": "7.9.0",
     "ts-node": "10.2.1",
-    "typescript": "4.3.5",
-    "uvu": "0.5.1",
-    "watchlist": "0.2.3"
+    "typescript": "4.4.3"
+  },
+  "ava": {
+    "files": [
+      "test/**/*.spec.ts"
+    ],
+    "extensions": {
+      "ts": "module"
+    },
+    "nodeArguments": [
+      "--loader=ts-node/esm"
+    ]
   },
   "workspaces": [
     "demo"
